Refactor `project_on_boundary` into a single file
